sorry i forgot i did

+++ src/controllers/videoController.js
@@ -1,7 +1,10 @@
 import Video, { formatHashtags } from "../models/Video";
+import User from "../models/User";
 
 export const home = async (req, res) => {
-  const videos = await Video.find().sort({ createdAt: "desc" });
+  const videos = await Video.find()
+    .sort({ createdAt: "desc" })
+    .populate("owner");
   return res.render("home", { pageTitle: "Home", videos });
 };
 export const search = async (req, res) => {
@@ -10,13 +13,14 @@ export const search = async (req, res) => {
   if (keyword) {
     videos = await Video.find({
       title: { $regex: new RegExp(keyword, "i") },
-    });
+      hashtags: { $regex: new RegExp(keyword, "i") },
+    }).populate("owner");
   }
   return res.render("search", { pageTitle: "Search", videos });
 };
 export const watch = async (req, res) => {
   const { id } = req.params;
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate("owner");
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Page not found." });
   } else {
@@ -25,24 +29,50 @@ export const watch = async (req, res) => {
 };
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  console.log(id);
+  const {
+    user: { _id },
+  } = req.session;
+  const video = await Video.findById(id);
+  const user = await User.findById(video.owner);
+  if (!video) {
+    res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
+  user.videos.splice(user.videos.indexOf(id), 1);
+  user.save();
   await Video.findByIdAndDelete(id);
   res.redirect("/");
 };
 
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id);
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Page not found." });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   return res.render("edit", { pageTitle: `Edit: ${video.title}`, video });
 };
 export const postEdit = async (req, res) => {
   const { id } = req.params;
-  const video = await Video.exists({ id: id });
+  const {
+    user: { _id },
+  } = req.session;
+  const video = await Video.findById(id);
   const { title, description, hashtags } = req.body;
   if (!video) {
     return res.render("404", { pageTitle: "Page not found." });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   try {
     await Video.findByIdAndUpdate(id, {
@@ -52,6 +82,7 @@ export const postEdit = async (req, res) => {
     });
   } catch (error) {
     console.log(error);
+    return res.status(403).redirect("edit");
   }
   return res.redirect(`/videos/${id}`);
 };
@@ -60,18 +91,40 @@ export const getUpload = (req, res) => {
   return res.render("upload", { pageTitle: "Upload Video" });
 };
 export const postUpload = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
+  const { path: fileUrl } = req.file;
   const { title, description, hashtags } = req.body;
 
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       title,
+      fileUrl,
+      owner: _id,
       description,
       hashtags: Video.formatHashtags(hashtags),
     });
+    const user = await User.findById(_id);
+
+    user.videos.push(newVideo._id);
+    user.save();
+
     return res.redirect("/");
   } catch (error) {
     return res
       .status(400)
       .render("upload", { pageTitle: "Upload", errorMessage: error.message });
   }
+};
+
+export const registerView = async (req, res) => {
+  const { id } = req.params;
+  const video = await Video.findById(id);
+  if (!video) {
+    return res.sendStatus(404);
+  }
+  video.meta.views = video.meta.views + 1;
+  await video.save();
+  return res.sendStatus(200);
 };
